renaming, new lib, and other
renamed poker-main to projecteuler52-main. separated out the poker specific functions into new pokerlib module. wrote getHandRank function. other changes as well.
