+++ projecteuler54-main.py
